added review model

+++ src/app/api/campgrounds/[id]/route.ts
@@ -1,21 +1,28 @@
 import dbConnect from "@/lib/dbConnect";
 import Campground from "@/models/Campground";
+import { getServerSession } from "next-auth";
 import { NextRequest } from "next/server";
+import { options } from "../../auth/[...nextauth]/options";
 
 export async function GET(req: NextRequest, {params}: {params: {id: string}}) {
     try {
         await dbConnect();
-        const campground = await Campground.findById(params.id).populate('owner');
+        const campground = await Campground.findById(params.id).populate('owner').populate('reviews');
         if (!campground) {
             return Response.json({success: false, message: "Campground Not Found"});
         }
         return Response.json({success: true, data: campground});
     } catch (error) {
+        console.log(error);
         return Response.json({success: false, message: "Error fetching data from Database"});
     }
 }
 
 export async function PUT(req: NextRequest, {params}: {params: {id: string}}) {
+    const session = await getServerSession(options);
+    if (!session) {
+        return Response.json({success: false, message: "You must login first!!"});
+    }
     const data = await req.json();
     try {
         await dbConnect();
@@ -27,6 +34,10 @@ export async function PUT(req: NextRequest, {params}: {params: {id: string}}) {
 }
 
 export async function DELETE(req: NextRequest, {params}: {params: {id: string}}) {
+    const session = await getServerSession(options);
+    if (!session) {
+        return Response.json({success: false, message: "You must login first!!"});
+    }
     try {
         await dbConnect();
         await Campground.findByIdAndDelete(params.id);
